lambda/requirements.txt
PyGithub
slack-sdk
typing_extensions